remove url_for calls to old download endpoints
