+++ sfa_dash/templates/data/aggregate.html
@@ -10,7 +10,6 @@
 {% if upload_link is defined %}
 <a role="button" class="btn btn-primary btn-sm" href="{{upload_link}}">Upload data</a>
 {% endif %}
-<a role="button" class="btn btn-primary btn-sm" href="{{ url_for('forms.download_aggregate_data', uuid=aggregate['aggregate_id']) }}">Download data</a>
 <a role="button" class="btn btn-danger btn-sm" href="{{ url_for('data_dashboard.delete_aggregate', uuid=aggregate['aggregate_id']) }}">Delete</a>
 
 <h3 class="mt-3">Observations</h3>
@@ -53,25 +52,7 @@
 {% if bokeh_script is defined %}
 {{ bokeh_script | safe }}
 {% endif %}
-<script src="/static/js/plot-bound-parsing.js"></script>
-<div class="form-group mt-3">
-<div class="form-element">
-  <label for="start">Start (UTC):</label><br/>
-  <input type="date" id="start-date" name="period-start-date" {% if form_data is defined %}value="{{form_data['period-start-date']}}"{% endif %}required>
-  <input type="time" id="start-time" name="period-start-time" {% if form_data is defined %}value="{{form_data['period-start-time']}}"{% else %} value="00:00"{% endif %} required>
-</div>
-<div class="form-element">
-  <label for="end">End (UTC)</label><br/>
-  <input type="date" id="end-date" name="period-end-date" {% if form_data is defined %}value="{{form_data['period-end-date']}}"{% endif %}required>
-  <input type="time" id="end-time" name="period-end-time" {% if form_data is defined %}value="{{form_data['period-end-time']}}"{% else %} value="00:00"{% endif %} required>
-</div>
-<div class="form-element">
-  <button type="submit" form="plot-range-adjust" value="Submit" class="btn btn-primary">Set time range</button>
-</div>
-<form action="" method="get" id="plot-range-adjust" onsubmit="ParseStartEnd()">
-<input type="text" id="start" name="start" hidden>
-<input type="text" id="end" name="end" hidden>
-</div>
 {% endif %}
+{% include "data/time_widgets.html" %}
 {% endif %}
 {% endblock %}
