+++ sfa_dash/templates/data/time_widgets.html
@@ -1,0 +1,31 @@
+{% import "forms/form_macros.jinja" as form %}
+{# Time widgets are used for either setting the bounds of the plot or download #}
+<div class="form-group mt-3">
+<div class="form-element">
+  <label for="start">Start (UTC):</label><br/>
+  <input type="date" id="start-date" name="period-start-date" value="{{ period_start_date }}" required>
+  <input type="time" id="start-time" name="period-start-time" value="{{ period_start_time }}" required>
+</div>
+<div class="form-element">
+  <label for="end">End (UTC)</label><br/>
+  <input type="date" id="end-date" name="period-end-date" value="{{ period_end_date }}" required>
+  <input type="time" id="end-time" name="period-end-time" value="{{ period_end_time }}" required>
+</div>
+{# form makes a post request to download data #}
+<div class="form-element full-width">
+  <button type="submit" form="plot-range-adjust" value="Submit" class="btn btn-primary">Set time range</button>
+  <form action="" method="post" id="download-form" onsubmit="ParseStartEnd()">
+  <button type="submit" form="download-form" value="Submit" class="btn btn-primary">Download data</button>
+  <label for="format">Format:</label>
+  <input type="radio" name="format" value="text/csv" checked>CSV</input><a href="https://solarforecastarbiter.org/datamodel/#csv" target="_blank">&#x1f6c8;</a>
+  <input type="radio" name="format" value="application/json">JSON</input><a href="https://solarforecastarbiter.org/datamodel/#json">&#x1f6c8;</a>
+  <input type="text" class="start" name="start" hidden>
+  <input type="text" class="end" name="end" hidden>
+  {{ form.token() }}
+  </form>
+</div>
+{# Form reloads the page with start and end query parameters #}
+<form action="" method="get" id="plot-range-adjust" onsubmit="ParseStartEnd()">
+<input type="text" class="start" name="start" hidden>
+<input type="text" class="end" name="end" hidden>
+<script src="/static/js/plot-bound-parsing.js"></script>
